clang-tidy fix

+++ modules/footpaths/src/transfer_requests.cc
@@ -55,7 +55,8 @@ std::vector<transfer_requests> build_transfer_requests(
   LOG(info) << "Generated " << result.size() << " transfer requests.";
   LOG(info) << "Found " << targets << " targets in total.";
   LOG(info) << "Identified "
-            << (double)((double)targets / (double)result.size())
+            << (static_cast<double>(targets) /
+                static_cast<double>(result.size()))
             << " targets per source.";
   LOG(info) << "Found " << no_targets
             << " (src, profile)-tuples w/o targets. (not included in transfer "
